+++ packages/core/data-transfer/lib/providers/local-strapi-source-provider/__tests__/entities.test.ts
@@ -2,7 +2,12 @@ import type { IEntity } from '../../../../types';
 
 import { Readable, PassThrough } from 'stream';
 
-import { collect, getStrapiFactory, getContentTypes, createMockedQueryBuilder } from './test-utils';
+import {
+  collect,
+  getStrapiFactory,
+  getContentTypes,
+  createMockedQueryBuilder,
+} from '../../../__tests__/test-utils';
 import { createEntitiesStream, createEntitiesTransformStream } from '../entities';
 
 describe('Local Strapi Source Provider - Entities Streaming', () => {
